(12/04) Proteção de rotas no front-end e no back-end

+++ back-end/src/controllers/users.js
@@ -7,13 +7,16 @@ const controller = {}     // Objeto vazio
 controller.create = async function(req, res) {
   try {
 
-  
-    //Verifica se existe o campo password e o 
-    //criptografa antes de criar o novo usario
-    if(req.body.password){
+    //Somente usuarios administradores podem acessar este recurso
+    //HTTP 403: Forbinde
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
+    // Verifica se existe o campo "password" e o
+    // criptografa antes de criar o novo usuário
+    if(req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12)
     }
-    
+
     await prisma.user.create({ data: req.body })
 
     // HTTP 201: Created
@@ -29,9 +32,15 @@ controller.create = async function(req, res) {
 
 controller.retrieveAll = async function(req, res) {
   try {
+    //Somente usuarios administradores podem acessar este recurso
+    //HTTP 403: Forbinde
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
+
     const result = await prisma.user.findMany(
-      //Omite o campo "password"do resultado por questão de segurança
-      {omit: {password:true}}
+      // Omite o campo "password" do resultado
+      // por questão de segurança
+      { omit: { password: true } }  
     )
 
     // HTTP 200: OK (implícito)
@@ -47,8 +56,15 @@ controller.retrieveAll = async function(req, res) {
 
 controller.retrieveOne = async function(req, res) {
   try {
+
+    //Somente usuarios administradores podem acessar este recurso
+    //HTTP 403: Forbinde
+    if(! (req?.authUser?.is_admin)|| Number(req?.authUser?.id) === Number(req.params.id)) return res.status(403).end()
+
     const result = await prisma.user.findUnique({
-      omit: {password:true},
+      // Omite o campo "password" do resultado
+      // por questão de segurança
+      omit: { password: true },
       where: { id: Number(req.params.id) }
     })
 
@@ -68,9 +84,14 @@ controller.retrieveOne = async function(req, res) {
 controller.update = async function(req, res) {
   try {
 
-    //Verifica se existe o campo password e o 
-    //criptografa antes de criar o novo usario
-    if(req.body.password){
+    //Somente usuarios administradores podem acessar este recurso
+    //HTTP 403: Forbinde
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
+
+    // Verifica se existe o campo "password" e o
+    // criptografa antes de criar o novo usuário
+    if(req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 12)
     }
 
@@ -94,6 +115,11 @@ controller.update = async function(req, res) {
 
 controller.delete = async function(req, res) {
   try {
+
+    //Somente usuarios administradores podem acessar este recurso
+    //HTTP 403: Forbinde
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
@@ -136,22 +162,23 @@ controller.login = async function(req, res) {
 
       // Usuário encontrado, vamos conferir a senha
       let passwordIsValid
-
-      //REMOVENDO VULNERABILIDADE DE AUTENTICAÇÃO FIXA
-      //if(req.body?.username === 'admin' && req.body?.password === 'admin123') passwordIsValid = true
-      //else
-      //passwordIsValid = user.password === req.body?.password
       
-      //Chamando bcrypt.compare() para verificar se o hash da senha
-      //enviada coincide com hash da senha armazenada no BD
+      // REMOVENDO VULNERABILIDADE DE AUTENTICAÇÃO FIXA
+      // if(req.body?.username === 'admin' && req.body?.password === 'admin123') passwordIsValid = true
+      // else passwordIsValid = user.password === req.body?.password
+      // passwordIsValid = user.password === req.body?.password
+      
+      // Chamando bcrypt.compare() para verificar se o hash da senha
+      // enviada coincide com o hash da senha armazenada no BD
       passwordIsValid = await bcrypt.compare(req.body?.password, user.password)
-
 
       // Se a senha estiver errada, retorna
       // HTTP 401: Unauthorized
       if(! passwordIsValid) return res.status(401).end()
-     //Remove o campo "password" do objeto user antes de usa-lo no token de resposta da requisição
-      delete user.password   
+
+      // Remove o campo "password" do objeto "user" antes
+      // de usá-lo no token e na resposta da requisição
+      delete user.password
 
       // Usuário e senha OK, passamos ao procedimento de gerar o token
       const token = jwt.sign(
@@ -171,7 +198,7 @@ controller.login = async function(req, res) {
 
       // Retorna o token e o usuário autenticado com
       // HTTP 200: OK (implícito)
-      res.send({token, user})
+      res.send({user})
 
   }
   catch(error) {
@@ -180,6 +207,13 @@ controller.login = async function(req, res) {
     // HTTP 500: Internal Server Error
     res.status(500).end()
   }
+}
+
+controller.logout = function(req, res){
+  //Apaga no front-end o cookie que armazena o token
+  res.clearCookie(process.env.AUTH_COOKIE_NAME)
+  //HTTP 204: No content
+  res.status(204).end()
 }
 
 controller.me = function(req, res) {
